changed floAT TO INT

+++ main.py
@@ -5,7 +5,7 @@ def digit_sum(n):
     return n % 10 + digit_sum(n// 10)
 
 def run():
-  n=float(input("Enter an int: "))
+  n=int(input("Enter an int: "))
   print(f"sum of digits of {n} is {digit_sum(n)}.")
 
 run()
